backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Security
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.8.0

# Elasticsearch
elasticsearch==8.11.0

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# HTTP client
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Code quality
ruff==0.1.6
mypy==1.7.1
black==23.11.0

# Utilities
python-dotenv==1.0.0
